feature: Support for non str envvars
